cleaned up code into components
